fix(reporting): TAMOC-381: Moved connection logic in central database back to index file
