+++ packages/central-server/app/database/index.js
@@ -1,1 +1,41 @@
-export { initDatabase, closeDatabase } from './connection';
+import config from 'config';
+
+import { addHooks } from './hooks';
+import { closeAllDatabases, openDatabase } from '@tamanu/database/services/database';
+import { setFhirRefreshTriggers } from './setFhirRefreshTriggers';
+
+const getOrCreateConnection = async ({ testMode, ...configOverrides }, key = 'main') => {
+  const store = await openDatabase(key, {
+    ...config.db,
+    ...configOverrides,
+    testMode,
+  });
+
+  // drop and recreate db
+  if (testMode) {
+    await store.sequelize.drop({ cascade: true });
+    await store.sequelize.migrate('up');
+  }
+  if (key === 'main') {
+    await addHooks(store);
+  }
+
+  await setFhirRefreshTriggers(store.sequelize);
+
+  return store;
+};
+
+export async function initDatabase({ testMode = false, dbKey = 'main' }) {
+  // connect to database
+  return await getOrCreateConnection(
+    {
+      testMode,
+      saltRounds: config.auth.saltRounds,
+    },
+    dbKey,
+  );
+}
+
+export async function closeDatabase() {
+  return closeAllDatabases();
+}
